Improve caching of Twinkle preferences
Currently, twinkleoptions.js is fetched on every page load. If the options haven't been edited since the last fetch, the response is typically 304 and the data isn't refetched, but the request is still necessary to check if the there has been any change.

This commit switches the fetch to use the API instead of index.php?action=raw. The maxage/smaxage parameters are used in the API call, which causes the response to be cached in the browser in a way that avoids revalidation requests on every fetch. This makes it significantly faster than requests that give 304. Expiry of 1 day (86400 seconds) is used for the HTTP cache. Hard reload can be done to clear the cache without waiting for expiry.

Since Twinkle initialization waits for preferences to be available, this also improves Twinkle's load performance.

Closes #2185